In this exercice we will implement a game. This game consist of Characters and Weapons.
Each character have:

health points between 0 and 100, 100 means the character’s health is full, 0 means the character is dead.
default damage points, when a character does not have a weapon, this is the default damage points to use when attacking.
weapon
attack: a method used to attack other character.
defend: a method used when being attacked.
Each weapon have:

damage points: when attacking the amount of points to remove from a player’s health.
PS: feel free to choose the different values for characters and weapons.